v0.0.14] edit : Obstacle Base Distance

+++ Assets/_Scripts/GameManager.cs
@@ -2,5 +2,5 @@ public static class GameManager
 {
     public static bool IsPlaying = false;
 
-    public static float mileage = 0f;
+    public static double mileage = 0f;
 }
